narrowed down culprits to 5 files
